+++ src/py-interface/settings_win.py
@@ -5,11 +5,12 @@ import gobject
 import gtk
 
 from _window_form import _WindowForm
+import agn_binder as ab
 
 class ConfigurationWindow(_WindowForm):
     """ConfigurationWindow"""
 
-    def __init__(self, values=False):
+    def __init__(self, vpn, values, change_password_win):
         super(ConfigurationWindow, self).__init__()
 
         self.set_resizable(False)
@@ -34,10 +35,18 @@ class ConfigurationWindow(_WindowForm):
         close_save.connect('clicked', self.do_btn_save)
         self._attach(close_save, 1, 3)
 
+        self.change_password = button = gtk.Button(_('Change Password'))
+        button.set_tooltip_text(_('You should disconnect first'))
+        button.connect('clicked', present_window(change_password_win))
+        self._attach(button, 2, 3)
+
         self.table.show_all()
         self.connect('form_submit', self.__on_submit__)
-        if values:
-            self.set_values(values)
+
+        vpn.connect('agn_state_change', self.on_agn_state_change)
+        self.on_agn_state_change(vpn, vpn.get_state())
+
+        self.set_values(values)
 
     def set_values(self, values):
         """set_values"""
@@ -56,12 +65,20 @@ class ConfigurationWindow(_WindowForm):
                           self.txt_username.get_text(),
                           self.txt_password.get_text())
 
+    def on_agn_state_change(self, vpn, state):
+        enabled = state < ab.STATE_BEFORE_CONNECT
+        self.change_password.set_sensitive(enabled)
+        self.change_password.set_has_tooltip(not enabled)
+
     def on_password_visiblity(self, widget):
         """on_password_visiblity"""
         self.txt_password.set_visibility(widget.get_active())
 
     def __on_submit__(self, _):
         self.do_btn_save(False)
+
+def present_window(win):
+    return lambda w: win.present()
 
 gobject.signal_new('save', ConfigurationWindow, gobject.SIGNAL_RUN_FIRST,
     None, (str, str, str, ))